pass 2nd stage

+++ src/Main.cpp
@@ -23,7 +23,7 @@ json decode_bencoded_value(const std::string& encoded_value) {
         }
     } else if (encoded_value[0]=='i') {
         std::string number_string = encoded_value.substr(1, encoded_value.find('e'));
-        int number = stoi(number_string);
+        int64_t number = stoll(number_string);
         return json(number);
     }
     else {
